Write as much as possible to the runner's Writer
Once we have an fully initialized runner we should write as much as
possible to the runner's internal Writer, including any errors that may
occur from creating, and executing jobs on a Driver.

The runner Job API has been updated too. The Fail method now accepts an
error, when not nil this will be appended to the Job's internal Errors
slices, which is then iterated over at the end of a run for displaying
any Driver runtime errors.

The Docker Driver now builds up a shell string to be run on the
container, similar to how the QEMU Driver builds up a shell string.