Allow for custom Job buffers
Previously we buffered up a Driver's output then just copied it over to
the Runner's output. This was inefficient and would give the impression
of builds hanging as there was little visual feedback.

I think I will do away with using a ReadWriter on the Job struct
entirely, and just use a Writer instead. It's an interface so the user
can implement their own if they'd like to keep a copy of whatever is
written to a job.

Right now this code is not thread-safe, and could lead to mangled output
being written to Stdout from the runner.