Clean up source URLs
Source URLs are now specified as a simple string that get's parsed by
the YAML unmarshaler. They follow the below format:

  [url] [ref] => [dir]

The [ref] and [dir] parts are optional. If not presented then the [ref]
will be set to 'master' and the 'dir' will be set to the base of the
given URL.