Begin work on build objects
Build objects are files that are passed from the host to the build
environment, essentially the opposite of an artifact. These can be
useful in situations if you need to have access to files that would
otherwise not sit in version control, and therefore not be pulled in
from one of the specified sources.

Right now the plan is for objects to only be supported in the offline
runner, I need to figure out a good way of having objects supported on
the server.

Under the hood the config.Artifact type has been renamed to
config.Passthrough, since that better captures what is happening.

The plan is to allow both files, and directories to be passed between
host and driver, this will require some revamping of how artifact
collection works, along with how object placment works.