Implement Placer interface for placement of objects
Both the Collector and Placer interfaces are used for file passthrough
between the host and the driver. This allows for flexibility, for
example you could have the Placer interface pull from an FTP server, if
such an implementation exists, and place files on the driver, and have
the Collector interface collector to the local filesystem.