Implement object passthrough for Docker driver
Right now this is only a prototype. We rely on os.Stat to get
information about the object we're placing. For now, this works. Only
because we have a Filesystem placer. If we were to introduce an FTP
placer for example, then this wouldn't work, we need a more reliable way
of writing TAR headers from the object's source, perhaps implement a
Stat method on the Placer interface? Needs investigation.