Refactor model logic - remove DB package variable
Each model now requires a direct connection to the database, so they no
longer rely on the DB package level variable.

Stores are now used for retrieving models from the database. These will
set the database connection on each model retrieved from the database so
the model can effectively be used. Models can also return stores for
related model data. For example, a User model can return a BuildStore,
which would allow for querying builds for that user.

I took this approach because I wasn't particularly fond of relying on a
package level variable for interacting with the database. This means the
web handlers will need a direct dependency on the stores they want to
use, which is more preferred.