Don't encode model fields in forms, remove init function
The model fields in the form structs are not required for flashing the
old form data to the session. Instead a bind method will be added to
each form, should that form require a model to be bound to it for
querying the database.

The 'init' function in the 'form' package has been removed. The gob
registration is now done as part of the UI server initialization
instead.