Refactor collector and placer into FileStore interface
This new interface will implement both runner.Collector and
runner.Placer interfaces, as well as provide a way of performing common
operations on files such as opening, creating, and removing.

This can allow for multiple implementations of artifact, and object
filestores, such as the local file system or an FTP server.