Refactor model package to use query builder
This query builder uses self-referential functions for building SQL
queries in an idiomatic way to Go. These queries then get handed off to
sqlx for the unmarshalling of the database data.

This allows for a more programmatic approach to building queries, and
allows for already existing queries to be easily extended.