Refactor driver configuration for server and worker
Instead of having a separate queue for each driver, a single queue will
be used instead. The name of this queue will contain the name of the
drivers supported by the server/worker. For example if qemu and docker
are configured then the queue name would be 'thrall_builds_docker_qemu'.
The driver names will be sorted alphabetically, so ordering of the
driver names in the config files will not matter.