Begin work in retrieving provider repositories for integration
Repositories will be cached in redis for 1 hour. Each provider will have
a corresponding function that is called to simply load in the
repositories for the user. I don't anticipate needing anymore
functionality from the providers such as GitHub or GitLab which is why
this simple approach is being taken.