Make runner.Passthrough a struct with a map property
This has been done as the go-yaml library would previously unmarshal this type
as a map instead of calling the MarshalYAML method on the type. Now this
can be properly unmarshalled into a string slice. This also allows for
us to directly use the config.Manifest type on the model.Build type
directly.