Use ToggleRepo method to toggle repository hooks
We now keep track of the hook ID we receive from the provider API so we
can delete it upon the repository being disabled. Along with this, the
name, or href is no longer stored in the database, since this is
information we get directly from the API. We're only interested in
storing the IDs for the repository, and the hook itself.