Implement runner.Placer and runner.Collector using stores
The stores have access to the database, so it makes sense to implement
the necessary interfaces for the runner.Runner here. They can hit the
database to see if the record exists, and to get the correct file hash
to use when calling the underlying runner.Placer/runner.Collector
interface respectively.