Refactor thrall-worker
A custom runner.Placer implementation is now used, this can be used for
placing keys and objects into the worker's build environment, and works
much better than the previously hacky solution.

Since the entity stores are now implementations of the runner.Placer,
and runner.Collector interfaces we can just pass these through to the
runner created by the worker, instead of having to implement these in
the worker lack we did last time.