Refactor runner.Driver implementations
Callbacks are now using to configure each driver. The drivers should now
have their configuration functions registered against the runner too,
this should make instantiating the drivers when needed simpler. For
example, when the thrall-worker is starting up it would register the
driver configs there, and use this to determine if the drivers have been
properly configured, allowing for a fast-fail if they haven't been.