Finalize driver validation and configuration
The new config.ValidateDrivers function will validate the contents of
the thrall-driver.toml file to make sure all of the configured drivers
are using valid configuration options.

The config.GetDriverConfig function now replaces the old
runner.GetDriver function. This is because,

* The runner package shouldn't really care about how a driver is
configured, it should only care that it receives an implementation of
runner.Driver

* It makes more sense for this to be in the config package since drivers
are part configured by what's in the TOML file, and what is in the
manifest.

The new driver.ConfigureFunc type is what will allow drivers to be
configured individually. The global config will be in the TOML file,
such as QEMU disk location, SSH timeout etc. Whereas individual config
will be in the build manifest, such as the driver image to use.