Refactor namespace.ResourceForm into namespace.Resource
This no longer implements the form.Form interface as it doesn't need to.
Instead it supplies a BindNamespace method that will bind the namespace
for the current resource to the given model.Binder. The previous
implementation checked to see if the current user could access the
namespace, but didn't bother binding to the namespace, which is
beneficial when checking the uniqueness of a resource's name.