Keep order with slices, lookup artifact hash via name, handle drive create
Order of jobs and stages are now maintained via slices. This caused bugs
where jobs and stages would be executed out of order - not good.

During artifact collection the name of the artifact is used to lookup
the artifact's hash. This will stop the build log from being polluted
with the artifact hash, for readability.

A new callback has been added to runner.Runner for handling the driver
creation. This is now used to capture the time at which the driver began
creation. Now we can effectively capture the duration it takes to create
a driver for builds.