Move as much domain logic from the handlers into top-level packages
Big change here, basically the HTTP handlers themselves should be
treated as an I/O layer, and the domain logic should belong in the
top-level packages. For example build submission and creation now exist
in the build package, instead of in the build/handler package where it
previously was.

Also finish implementing the rest of the API, next up integration tests
to test the API.