Fix namespace logic bugs around ownership and collaborators
We no longer call SharedWith on calls to namespace.Store.Get, as this
was causing weird errors were non-collaborator users could access a
namespace if it shared the same name as another.

When a collaborator is added to a namespace, they will always be given
access to the root namespace. This makes things simpler.