Create a provider for each group a user is in
Working on supporting org repos with webhooks, this requires needing to
lookup the provider tokens using the org ID instead of the user ID.
Workaround now is to simply create a new provider for each org that user
will be in.

This may change depending on how GitLab handles these things.