Allow for namespace names to be unique per user
The namespace.Store does not apply database.Where on the bound User
model. This is because namespace's are a resource that can be accesses
from unauthenticated users if they are public, so applying this would
result in 404s.

However, this was also causing namespace names to be unique across the
entire database instead of on a per user basis. Now database.Where is
applied to the namespace queries executed within getFromOwnerPath.