Move namespace ownerhsip check to getFromOwnerPath
When looking up a namespace using the "<path>@<username>" syntax we want
to do the ownership checks there instead of outside. If the user looking
up the namespace cannot add to it then the GetByPath call will return
ErrPermission. If the namespace being looked up does not exist and is
being checked against a different user, then we should treat this as a
permission failure too.