Refactor component configuration and server bootstrapping
Configuration for the various components, server, worker, scheduler, and
curator has been refactored. Now the actual resources required for
configuring these components exists in the config structs themselves.
For example the config struct config.Server will have the connections to
the SQL database and Redis.

This now allows for a more graceful way of setting up components such as
the server. This is now done via the serverutil package. This will
simply take the config.Server struct and pass it around the various
routers we need for registering the server's routes.

We have yet to make the same refactor to the djinn-worker, but this will
be done. I expect a similar workerutil package to be created to
facilitate this, similar to what we have for the server.