Use mcmathja/curlyq instead of machinery
I find this library simpler to work with for a background jobs in Go.
Plus it's a benefit I can pass it a pre-existing Redis connection to the
consumers/producers.

This also includes a general refactor of some of the code, for example
the manifest package now exists for working with build manifests,
whereas this code previously existed in the config package. The reason
for this split is because I want the config package to be used for
configuring the components of Djinn and the manifest itself isn't
necessarily used for configuring an individual component, it is simply
used by multiple components.