Refactor worker bootstrapping logic into workerutil
Similar to what we did with serverutil, workerutil now exists for
bootstrapping the worker that is used for processing the builds in
Djinn. The worker package now exists too that contains the logic for
executing a build, updating its status in the database, and handling
object placement and artifact collection.