Use build numbers per user make resources owned by namespace owner
Build numbers are now used and are tied to a user. When a build is
submitted from a collaborator to a namespace then this build will
count towards the namespace owner's build count.

Namespace resources will now be put under the ownership of the namespace
too. So when a collaborator adds a resource to a namespace (object,
image, variable, etc.) then this will be uploaded as the namespace
owner's resource.