Only attempt a build once
A build would only need to be reattempted if there is an internal error
from the worker itself. This is a CI system, so it's safe to assume that
build manifests would be submitted on a regular basis. Should an
internal error occur on the worker (database connection issues) then
build will fail, but assuming build manifests are being submitted on a
frequent basis this should have a minimal impact.