Clean up how manifests are marshalled
When a manifest is marshalled into a string it will have a tendency of
clobbering defaults into what was submitted. For example, if no ref was
given then master would be in string form of the manifest, which is not
ideal since we want to use the default branch of a source repo, whatever
that may be.

This has now been remediated so there is less clobbering of manifest
sources.