Use a queue per driver unless otherwise specified
This need arose from the fact that we need to ensure that the qemu
driver uses the same architecture for the guest and the host. This way
we can utilize KVM as the hypervisor. We do this by specifying the full
arch to use for the qemu driver. For example to run a djinn-server for
qemu builds on amd64 you would put,

    drivers = ["qemu-x86_64"]

and for the worker you would put,

    driver = "qemu-x86_64"

from the user's perspective when submitted builds they don't need to do
anything new. Simply specifying "qemu" would be enough.

The djinn-worker also has an additional check in place to ensure that
the qemu driver, if specified, will match the host architecture. If the
djinn-worker has been configured to work for all drivers, like so,

    driver = "*"

then the default qemu driver it will use will be for the architecture
the djinn-worker is running on.