Support bzip2 compression of images
Large image files will take a long time to upload, since they can be
>1GB on average, which is a shame. However we can somewhat mitigate this
by allowing users to upload bzip2 compressed images. During the upload
these will be decompressed, which won't incur that significant of a hit
when running on decent hardware.

Also we do not partition the stores used for custom images, with the way
the image directory is meant to be structured per driver, this won't
work.