Remove driver registry
This is no longer needed. Workers are now configured on a per driver
basis, instead of previously where a single worker could process builds
with different drivers. This makes things simpler, and now during
decoding of the driver configuration we only return the initialization
function for said driver, along with its global config.