Make key namespaces immutable
Key is one of the few resources that can actually be edited. When a key
is being edited we now only allow for the config itself to be edited.
This removes the possibility of someone moving it to another namespace.
If you wish to do this, then you must create a new key.