Attempt to redial on RSET failure
Currently we are getting sporadic broken pipe errors when we send
emails, which is no good. To try and mitigate this we will redial the
client whenever a RSET command fails. If the subsequent redial fails,
then we give up, as this would most likely point to a more severe issue
with the SMTP connection.