Move curlyq consumer to config.Worker
This prevents a collision between the in memory queue.Queue and the
previous Queue field we had. We don't need the name of the queue the
worker is consuming from anyway, since this was previously used to
create the consumer. Instead, we will simply setup the consumer during
config unmarshalling, then pass that consumer onto the worker itself.