Orphan a build if it has already started
We're making the assumption here that if a build already has a valid
StartedAt property, then it was because the worker restarted, so we mark
the build as orphaned. This will simply reset the build's output and
status back to being that of a queued build.

This needs some testing, and perhaps some additional logic for handling
the reaping of any of the driver processes that may be left lingering
when an orphaned build does occur.