General clean up of logic, and bug fixes
DownloadURL now implements its own GobEncode/GobDecode methods for
correctly encoding/decoding the underlying URL. Trying to decoded the
gob to the embedded struct pointer results in a panic.

We also removed the type-assert on the image upload file, as this is now
handled by andrewpillar/webutil itself via the Remove method.