Update dist config
The driver name will be appended to the path. This is to support
partitioning, for example if a user uploads a custom image it would be
uploaded to /var/lib/djinn/image/<id>/<driver> where <id> is the ID of
the user, and <driver> is the name of the driver.