Gracefully handle server start failures
Previously, since the server was started in a goroutine, any errors that
occurred when trying to bind to an address would not exit out of the
program and leave it to hang. Now, a channel is used to signal that the
server failed to bind to an address. This allows for gracefully closing
any connections that may have been opened up by the server.