Prompt user for an email if one is not set on their account
There can be occasions when a user authenticates via 3rd party provider
in which we cannot obtain their email address during the regular flow.
Down the line, this will cause issues. So, we've added the CheckEmail
middleware which will prompt the user for their email address if one is
not set. This uses a temporary account token for authentication between
requests -- this is because when an account is created during the oauth
flow the password is randomly generated.