Remove unused abstractions and fix bugs
This was a much needed cleanup of the code base that was needed. We've
removed some unused abstractions such as the server.Router interface in
place of a simple RegisterX function for registering routes. Where X
will either be API or UI depending on the kind of routes being
registered.

Integration tests have been expanded, and general bugs have been fixed.