Render pushpin alongside pinned builds in UI
This change was pretty far reaching since the build/template.Index is
used in multiple places. This change also involved making the base
template have a Query field which will capture all of the URL query
values for the current URL, this offers more flexibility when there is
data in the URL query values we may want to use when rendering a page.